fix: centralize language-select in 1 component
